+++ s3_tar/s3_tar.py
@@ -19,7 +19,10 @@ class S3Tar:
                  cache_size=5,
                  save_metadata=False,
                  remove_keys=False,
+                 allow_dups=False,
                  session=boto3.session.Session()):
+
+        self.allow_dups = allow_dups
         self.source_bucket = source_bucket
         self.target_bucket = target_bucket
         if self.target_bucket is None:
@@ -61,7 +64,7 @@ class S3Tar:
         if self.cache_size is None or self.cache_size <= 0:
             raise ValueError("cache size must be 1 or larger")
 
-        self.s3 = _create_s3_client(session)
+        self.s3 = _create_s3_client(session, self.cache_size * 2)
 
     def tar(self):
         """Start the tar'ing process with what has been added
@@ -197,39 +200,39 @@ class S3Tar:
         """Started as a background job to keep adding files to the
         file cache to speed things along
         """
-        def _fetch(key):
+        def _fetch(item):
             while len(self.file_cache) >= self.cache_size:
                 # Hold here until more files are needed in the cache
                 time.sleep(0.1)
-
+            tar_member_name, key = item
             logger.debug("Adding to cache {}".format(key))
-            self._add_key_to_cache(key)
+            self._add_key_to_cache(tar_member_name, key)
 
         _threads(self.cache_size, self.all_keys, _fetch)
         self.keys_to_delete = self.all_keys.copy()
         self.all_keys = set()  # clear now that all have been processed
 
-    def _add_key_to_cache(self, key):
+    def _add_key_to_cache(self, tar_member_name, key):
         """Get the source of an s3 key (and its metadata if needed) and add
         it to the file cache
 
         Args:
+            tar_member_name (str): Filename and path of the file inside the tar
             key (str): the key to download form s3
         """
         if self.save_metadata is True:
-            metadata_io = self._get_tar_source_metadata(key)
+            metadata_io = self._get_tar_source_metadata(tar_member_name, key)
             if metadata_io is not None:
                 logger.debug("Adding metadata file to cache {}".format(key))
                 self.file_cache.append(metadata_io)
 
-        self.file_cache.append(
-            self._get_tar_source_data(key)
-        )
+        self.file_cache.append(self._get_tar_source_data(tar_member_name, key))
 
-    def _get_tar_source_data(self, key):
+    def _get_tar_source_data(self, tar_member_name, key):
         """Download source file and generate a tar from it
 
         Args:
+            tar_member_name (str): Filename and path of the file inside the tar
             key (str): File from s3 to download
 
         Returns:
@@ -237,19 +240,20 @@ class S3Tar:
         """
         source_key_io = self._download_source_file(key)
         source_tar_io = self._save_bytes_to_tar(
-            key.split('/')[-1],
+            tar_member_name,
             source_key_io,
             mode=self.mode,
         )
         source_key_io.close()  # Cleanup
         return source_tar_io
 
-    def _get_tar_source_metadata(self, key):
+    def _get_tar_source_metadata(self, tar_member_name, key):
         """Get metadata from the s3 file
         If a file has metadata then add it to a tar file
         If not, then return None
 
         Args:
+            tar_member_name (str): Filename and path of the file inside the tar
             key (str): S3 file to get metadata from
 
         Returns:
@@ -260,7 +264,7 @@ class S3Tar:
             return None
 
         source_metadata_tar_io = self._save_bytes_to_tar(
-            key.split('/')[-1] + '.metadata.json',
+            tar_member_name + '.metadata.json',
             source_metadata_io,
             mode=self.mode,
         )
@@ -326,49 +330,97 @@ class S3Tar:
 
         return source_tar_io
 
-    def add_files(self, prefix):
+    def _raise_if_dup(self, tar_member_name, key, key_list=None):
+        if key_list is None:
+            key_list = self.all_keys
+
+        if any((True for x in key_list
+                if tar_member_name == x[0] and key != x[1])):
+            if self.allow_dups is False:
+                raise ValueError(("Filename '{member_name}' for key '{key}'"
+                                  " already exists in the tar file."
+                                  " Set allow_dups to continue.")
+                                 .format(member_name=tar_member_name,
+                                         key=key))
+
+    def add_files(self, prefix, folder='', preserve_paths=False):
         """Add s3 files from a directory inside the source bucket
 
         self.all_keys gets added to directly
 
         Args:
             prefix (str): Folder path inside the source bucket
+            folder (str, optional): Folders to place the file inside
+                the tar file. Defaults to ''.
+            preserve_paths (bool, optional): Starting from the prefix, use
+                the path of the key in the tar file. Defaults to False.
         """
         def resp_to_filelist(resp):
-            return [(x['Key']) for x in resp['Contents']]
+            file_list = []
+            for x in resp['Contents']:
+                key = x['Key']
+                if key == prefix:
+                    continue
 
-        objects_list = set()
+                # Get the paths after the prefix, and before the file
+                if preserve_paths is True:
+                    tar_member_name = folder + key.replace(prefix, '')
+                else:
+                    tar_member_name = folder + key.split('/')[-1]
+
+                self._raise_if_dup(tar_member_name, key, key_list=self.all_keys)
+                self._raise_if_dup(tar_member_name, key, key_list=file_list)
+
+                file_list.append((tar_member_name, key))
+
+            return file_list
+
+        # needs to end with '/'
+        if folder != '' and not folder.endswith('/'):
+            folder += '/'
+
         logger.info("Gathering files from folder {}".format(prefix))
         resp = self.s3.list_objects_v2(Bucket=self.source_bucket, Prefix=prefix)
         if resp['KeyCount'] == 0:
             logger.warning("No files found in the prefix {}".format(prefix))
             return
-        objects_list |= set(resp_to_filelist(resp))
-        if prefix in objects_list:
-            objects_list.remove(prefix)  # Do not add the prefix as a file
-        logger.debug("Found {} objects so far...".format(len(objects_list)))
+
+        file_list = resp_to_filelist(resp)
+        self.all_keys |= set(file_list)
+        total_file_count = len(file_list)
+
+        logger.debug("Found {} objects so far...".format(total_file_count))
         while resp['IsTruncated']:
-            last_key = objects_list[-1]
+            last_key = file_list[-1][1]
             resp = self.s3.list_objects(
                 Bucket=self.source_bucket,
                 Prefix=prefix,
                 Marker=last_key,
             )
-            objects_list |= set(resp_to_filelist(resp))
-            logger.debug("Found {} objects so far...".format(len(objects_list)))
+            file_list = resp_to_filelist(resp)
+            self.all_keys |= set(file_list)
+            total_file_count += len(file_list)
+
+            logger.debug("Found {} objects so far...".format(total_file_count))
 
         logger.info("Found {} objects in the folder '{}'"
-                    .format(len(objects_list),
-                            prefix))
-        self.all_keys |= objects_list
+                    .format(total_file_count, prefix))
 
-    def add_file(self, key):
+    def add_file(self, key, folder=''):
         """Add a single file at a time to be tar'd
 
         self.all_keys gets added to directly
 
         Args:
             key (str): The full path to a single s3 file in the source bucket
+            folder (str, optional): Folders to place the file inside
+                the tar file. Defaults to ''.
         """
         # TODO make sure key exists, if not log error and do not break
-        self.all_keys.add(key)
+        if folder != '' and not folder.endswith('/'):
+            folder += '/'
+        tar_member_name = folder + key.split('/')[-1]
+
+        self._raise_if_dup(tar_member_name, key)
+
+        self.all_keys.add((tar_member_name, key))